extract combobox e2e coverage
